+++ docs/pages/pagination.mdx
@@ -1,0 +1,39 @@
+import SEO from '../components/SEO';
+import { Button } from '@sajari/react-components';
+
+<SEO title="Pagination" description="Pagination component is used for paging through paged result sets." />
+
+# Pagination
+
+Pagination component is used for paging through paged result sets.
+
+```js
+import { Pagination } from '@sajari/react-components';
+```
+
+## Usage
+
+```jsx
+function PaginationExample() {
+  const [page, setPage] = React.useState(1);
+
+  return (
+    <div style={{ display: 'flex', justifyContent: 'center' }}>
+      <Pagination totalResults={1000} pageSize={20} page={page} onChange={setPage} />
+    </div>
+  );
+}
+```
+
+## Props
+
+The Pagination component composes the `ButtonGroup` component so you can pass props for
+`ButtonGroup`.
+
+| Name           | Type       | Default | Description                                                               |
+| -------------- | ---------- | ------- | ------------------------------------------------------------------------- |
+| `pageCount`    | `number`   |         | The total number of pages.                                                |
+| `totalResults` | `number`   |         | The total number of results (used to calculate page count if not set).    |
+| `pageSize`     | `number`   |         | The number of results per page (used to calculate page count if not set). |
+| `page`         | `number`   |         | The current page.                                                         |
+| `onChange`     | `function` |         | The callback handler. The current page is passed as the only argument.    |
