+++ packages/components/src/Pagination/types.ts
