+++ packages/components/src/assets/icons/index.tsx
@@ -1,4 +1,6 @@
 import Check from './check.svg';
 import Star from './small-star.svg';
+import ChevronLeft from './chevron-left.svg';
+import ChevronRight from './chevron-right.svg';
 
-export { Check, Star };
+export { Check, Star, ChevronLeft, ChevronRight };
