+++ packages/components/src/index.ts
@@ -11,4 +11,5 @@ export { default as Swatch } from './Swatch';
 export { default as Rating } from './Rating';
 export { default as Image } from './Image';
 export { default as Text } from './Text';
+export { default as Pagination } from './Pagination';
 export { default as ThemeProvider } from './styles/theming';
