+++ packages/components/src/utils/number.ts
@@ -2,7 +2,7 @@
  * Returns a number whose value is limited to the given range.
  *
  * Example: limit the output of this computation to between 0 and 255
- * (x * 255).clamp(0, 255)
+ * clamp((x * 255), 0, 255)
  *
  * @param {Number} input
  * @param {Number} min The lower boundary of the output range
