--- src/components/ButtonGroup/types.ts
@@ -1,16 +1,0 @@
-import { PropsWithAs } from "../../types/component-as";
-
-export interface Props {
-  /** Should the children be attached to each other */
-  attached?: boolean;
-  /** Make children fullWidth */
-  fullWidth?: boolean;
-  /** Display children horizontally */
-  inline?: boolean;
-  /** Children */
-  children: React.ReactNode;
-}
-
-export type UseButtonGroupStylesParams = Props["inline"];
-
-export type ButtonGroupProps = PropsWithAs<Props>;
