--- src/components/Pagination/Pagination.mdx
@@ -1,27 +1,0 @@
----
-name: Pagination
-menu: Components
-route: /components/pagination
----
-
-import { Playground } from 'docz';
-import { ExampleContainer, categoryMultiFilter } from '../../docs/helpers.ts';
-import ThemeProvider from '../../styles/theming';
-import Pagination from '.';
-
-# Pagination
-
-## Usage
-
-<Playground>
-  {() => {
-    const [page, setPage] = React.useState(1);
-    return (
-      <ExampleContainer>
-        <ThemeProvider>
-          <Pagination totalResults={1000} pageSize={20} page={page} onChange={setPage} />
-        </ThemeProvider>
-      </ExampleContainer>
-    );
-  }}
-</Playground>
